requirements.txt
black==23.11.0
distro==1.9.0
groq==0.22.0
huggingface-hub==0.30.1
matplotlib==3.8.2
mypy==1.7.0
mypy-extensions==1.0.0
numpy==1.26.2
packaging==24.2
pandas==2.1.3
pillow==11.1.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
pytorch-crf==0.7.2
requests==2.32.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tokenizers==0.13.3
tqdm==4.66.1
transformers==4.26.0
torch==2.4.1
torch-tb-profiler==0.4.3
torchvision==0.19.1